Implemented adding goods to the warehouse

+++ KS.Entities/Stock.cs
@@ -2,11 +2,11 @@
 {
     public class Stock : BaseEntity
     {
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
 
         public Product Product { get; set; }
 
-        public long WarehouseId { get; set; }
+        public int WarehouseId { get; set; }
 
         public Warehouse Warehouse { get; set; }
